Build model demonstration
This commit adds another contributor that generates a Maven pom.xml
file. This gives us an opportunity to show the build abstraction that
share commonalities when it makes sense. A pom.xml file is also created
when a Gradle build is requested as there isn't any condition that
discards this contributor if necessary.